keyboard is done
change keyboard to have search instead of return is done.
just need to implement the long click edit or delete now
will need to add in an update method in the database file
probably just open a exercisepopup for edit, with the info that already
have in it, and then update it with the new info they put in.
will need a prior popup to let them choose between deleting and
updating.