exercise-small C, new workout-lots of C
using "actionbaractivity" for exercises is causing this.oncreate(null)
to crash, which is annoying cuz i want it need to find something else

also another crash with new workout
same error message, so it may be similar, but it also may be other
stuff, cuz i never call an oncreate over there
planned to send that array back and forth between the two, but depending
on what this bug fix is. may need to do something else.